Successful ViewAllUsers Page. Begin DexView
Successfully created a ViewAllUsers list Admin panel for the frontend. Haven't yet added Edit, Details, or Delete for User Admin Page.

Beginning work on DexHolder view. This page will allow a user to view some of their personal information as well as a list of the people that they have created.

Successfully created a DexHolderMiddleVM viewmodel on the backend