API- Tweaked & Added Endpoints for user checking
Tweaked Login2 to send Username as well. Also tweaked LoginModel and Login2 to accept a generic Input field which can check for both Username and Email. May be useful for later, if I want to tweak the login view on the frontend.

Duplicated Login2 to LoginDouble.

LoginDouble will be used to check the currently logged in user, which should prevent other users from reaching other user's pages, unless they are in an admin or moderator account.

LoginDouble requires both Username and Email to work, which will be sent via the Frontend's UserSessionService