EntryItems list view works!
Had to do a quick fix for a method in DbPersonRepository on the backend so that it can properly parse out if the criteria was a PersonId (integer), or a string nickname (which was what it was defaulting to). Accomplished this using a get VM method for now, but I'd like to do a more elegant fix in the future.