EntryItems C,R1,U working + note on Int64....
So, a lot of issues I was facing was due to difficulties in handling Int64/longs, which don't play well with TempData or ViewData. In the future, I plan to change those Ids to Guids instead.

Delete is the last one to make, then I can retool them for Social Media