Delete & EntryItems complete. ViewDex Bugfixed
Will be duplicating and retooling for social media next

Also, fixed a bug where the first person's details that you try to view caused a 404. Solved it by querying the API two to three more times if the initial call didn't yield results. Probably has something to do with the route operating off of LocalCounter, a calculated column.