Last Commit
Final changes that were made for the time period of  this project. Changes that were made during this round of commits were not graded, and is primarily to serve the purpose of the state of the project when I decided to switch gears to another subject's final assignments.

After this point, the project should be forked.